Moving webproxy handling to its own file
